tutorial/requirements.txt
# common
django
djangorestframework
pygments

# pip install -r requirements.txt
# -r requirements.txt

django-extensions